\name{print}
\alias{print.RealVAMS}
\alias{print.summary.RealVAMS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print
}
\description{
Prints names of elements in RealVAMS object.
}
\usage{
\method{print}{RealVAMS}(x, ...)
}
\arguments{
  \item{x}{
object of class RealVAMS
}
  \item{\dots}{
other arguments to be passed to summary
}
}