\name{REML_Rm}
\alias{REML_Rm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function
}
\description{
An internal function
}
\usage{
REML_Rm(invsqrtW_,JYp_,loopsize_, patternlength_,rownumber_,ybetas_,etahat_,
tempmatR_,JXpi_,JXpp_,JXpx_,JXpdim_,JZpi_,JZpp_,JZpx_,JZpdim_,betacov_,C12_)
}
\arguments{
  \item{invsqrtW_}{
an internal variable
}
  \item{JYp_}{
an internal variable
}
\item{loopsize_}{
an internal variable
}
\item{patternlength_}{
an internal variable
}
\item{rownumber_}{
an internal variable
}
\item{ybetas_}{
an internal variable
}
\item{etahat_}{
an internal variable
}
\item{tempmatR_}{
an internal variable
}
\item{JXpi_}{
an internal variable
}
\item{JXpp_}{
an internal variable
}
\item{JXpx_}{
an internal variable
}
\item{JXpdim_}{
an internal variable
}
\item{JZpi_}{
an internal variable
}
\item{JZpp_}{
an internal variable
}
\item{JZpx_}{
an internal variable
}
\item{JZpdim_}{
an internal variable
}
\item{betacov_}{
an internal variable
}
\item{C12_}{
an internal variable
}
}