\name{vp_cp}
\alias{vp_cp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function
}
\description{
An internal function
}
\usage{
vp_cp(Z_mat, B.mat, control)
}
\arguments{
  \item{Z_mat}{
data frame
}
  \item{B.mat}{
data frame
}
\item{control}{
a list
}
}